refactor: split inner classes to seperate classes
